Moved unit tests to folder
